Rename `edit` to `document`
