+++ crates/server/src/document/text_diff.rs
