+++ crates/server/src/document/text_edit.rs
